requirements.txt
alembic
APScheduler
cryptography
email-validator==2.2.0
firebase-admin
Flask==3.1.1
Flask-CORS==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-migrate
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
geopy
google-api-python-client
google-auth
google-auth-oauthlib
gunicorn==23.0.0
Jinja2==3.1.6
openpyxl
Pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv
pytz
reportlab==4.4.3
requests==2.32.4
resend
sendgrid
SQLAlchemy==2.0.43
weasyprint==66.0
Werkzeug==3.1.3
WTForms==3.2.1
yagmail
alembic
APScheduler
cryptography
email-validator
firebase-admin
Flask
Flask-CORS
Flask-Login
Flask-Mail
flask-migrate
Flask-SQLAlchemy
Flask-WTF
geopy
google-api-python-client
google-auth
google-auth-oauthlib
gunicorn
Jinja2
openpyxl
Pillow
psycopg2-binary
PyJWT
python-dotenv
pytz
reportlab
requests
resend
sendgrid
SQLAlchemy
weasyprint
Werkzeug
WTForms
yagmail
